backend/requirements.txt
asgiref==3.2.10
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
generics==1.1.0
importlib-metadata==1.7.0
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.6
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
zipp==3.1.0
